requirements.txt
fastapi
uvicorn
python-docx
requests
python-multipart